+++ src/DataVoyager.jl
@@ -1,18 +1,27 @@
 __precompile__()
 module DataVoyager
 
-using Blink, DataValues
+using Electron, DataValues
 
 import IteratorInterfaceExtensions, TableTraits, IterableTables, JSON
 
 export Voyager
 
-mutable struct Voyager
-    w
-    function Voyager()
-        w = Blink.Window()
+app = nothing
 
-        loadfile(w, joinpath(@__DIR__, "htmlui", "main.html"))
+mutable struct Voyager
+    w::Window
+
+    function Voyager()
+        main_html_uri = string("file:///", replace(joinpath(@__DIR__, "htmlui", "main.html"), '\\', '/'))
+
+        global app
+
+        if app==nothing
+            app = Application()
+        end
+
+        w = Window(app, URI(main_html_uri), options=Dict("title"=>"Data Voyager"))
 
         new(w)
     end
@@ -29,12 +38,28 @@ function (v::Voyager)(source)
 
     data = JSON.json(data_dict)
 
-    jsdata = Blink.JSString(data)
+    code = "voyagerInstance.updateData($data)"
 
-    @js v.w voyagerInstance.updateData($jsdata)
+    run(v.w, code)
 
     return nothing
 end
+
+# function Base.getindex(v::Voyager)
+#     code = "voyagerInstance.getSpec(true)"
+
+#     content = run(v.w, code)
+
+#     return VegaLite.VLSpec{:plot}(content)
+# end
+
+# function Base.getindex(v::Voyager, index::Int)
+#     code = "voyagerInstance.getBookmarkedSpecs()"
+
+#     content = run(v.w, code)
+
+#     return VegaLite.VLSpec{:plot}(JSON.parse(content[index]))
+# end
 
 function Voyager(source)
     v = Voyager()