+++ NEWS.md
@@ -1,3 +1,6 @@
+# DataVoyager.jl v0.1.0 Release Notes
+* Use Electron.jl instead of Blink.jl
+
 # DataVoyager.jl v0.0.3 Release Notes
 * Bug fixes